+++ packages/arb-validator/ethbridge/messagesChallenge.go
@@ -188,10 +188,10 @@ func (c *messagesChallenge) ChooseSegment(
 	segmentHashes []common.Hash,
 	chainLength *big.Int,
 ) error {
-	bisectionCount := uint32(len(chainHashes) - 1)
+	bisectionCount := uint64(len(chainHashes) - 1)
 	bisectionHashes := make([]common.Hash, 0, bisectionCount)
-	for i := uint64(0); i < uint64(bisectionCount); i++ {
-		stepCount := structures.CalculateBisectionStepCount(i, uint64(bisectionCount), chainLength.Uint64())
+	for i := uint64(0); i < bisectionCount; i++ {
+		stepCount := structures.CalculateBisectionStepCount(i, bisectionCount, chainLength.Uint64())
 		bisectionHashes = append(
 			bisectionHashes,
 			structures.MessageChallengeDataHash(
